kl.ml: overhaul lexer ~ runs orders of magnitudes faster
previously, the lexer would be horrendously slow on medium sized files
(see t-star.kl from upstream Shen). The total amount of time required to
*just* lex the file was over a second. Running the code under callgrind
showed that over 99% of the time was spent in calling `List.length`.

So, to ameliorate this, I've introduced a new type `sized_list`, which
is identical to the standard list, except that it stores it's size with
the list itself as an int.

The reduction in time (to parse t-star.kl) is as follows:

`ocamlopt kl.ml -o kl`

Old: 1.37s
New: 0.00519s

A few orders of magnitude!